Prevent hangs in mxs173_throttle_filter
The test now sets a two minute timeout for all larger operations. This
prevents excessive waiting when the test is executed.

Removed the row count output to stdout to prevent excessive logging when
the terminal contents are written to a file.

Also renamed the file to match the test case name. This should remove the
need to find the source file to test name mapping from the CMakeLists.txt.